refactor(api): move `/payment/*` routes to own file
